requirements.txt
google-genai>=1.0.0
PySide6>=6.6.0
python-dotenv>=1.0.0
